allineamento corretto

+++ src/main/java/BusinessLogic/UserActionsController.java
@@ -1,10 +1,15 @@
 package main.java.BusinessLogic;
 
-import javafx.scene.control.Alert;
+import javafx.event.ActionEvent;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.control.*;
+import main.FXML.GUIControl.MessagesController;
+import main.FXML.GUIControl.SelectGuestsPaneController;
 import main.java.DomainModel.*;
 
 import main.java.ORM.*;
 
+import java.io.IOException;
 import java.sql.SQLException;
 import java.sql.Time;
 import java.sql.Date;
@@ -12,6 +17,7 @@ import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.ArrayList;
+import java.util.Optional;
 
 import static main.java.DomainModel.NotificationType.DELETION;
 import static main.java.DomainModel.NotificationType.MODIFICATION;
@@ -42,7 +48,7 @@ public class UserActionsController extends PersonController{
     }
 
     //methods
-    //FIXME check input parameters
+    //TODO it should be removed? Maybe yes
     public float calculatePricePerPerson(int idField, int nPeople) throws SQLException, ClassNotFoundException {
 
         FieldDao fieldDao = new FieldDao();
@@ -88,6 +94,7 @@ public class UserActionsController extends PersonController{
         return invitablePlayers;
     }
 
+    //TODO should be changed output type into boolean for manage success or error banner by caller?
     public void addReservation(Date eventDate, Time eventTimeStart, Time eventTimeEnd, Field field, int guests, int requiredParticipants, boolean isMatched, ArrayList<String> accounts) throws SQLException, ClassNotFoundException {
 
         ReservationDao reservationDao = new ReservationDao();
@@ -102,31 +109,27 @@ public class UserActionsController extends PersonController{
         reservation.setId(newReservationId); //WARNING: it's very important
 
         //group creation
-        Group group = new Group(user,reservation, requiredParticipants); //TODO check if participants and users array will be filled. Check constructor.
+        Group group = new Group(user,reservation, requiredParticipants);
         int newGroupId = groupDao.addGroup(group);
+        group.setId(newGroupId); //WARNING: it's very important
         joinGroup(newGroupId,guests);
 
         if (isMatched) {
             sendInvites(group);
-            //TODO add matchmaking and send invite methods
         }
-        else {
-            //TODO optimize it, if needed
-            UserDAO userDAO = new UserDAO();
 
-            for (String accountUsername : accounts){
-                sendInvite(reservation,userDAO.getUserID(accountUsername)); //TODO could be better by username than by id?
+        UserDAO userDAO = new UserDAO();
+
+        for (String accountUsername : accounts){
+            if (accountUsername != null) {
+                sendInvite(reservation, userDAO.getUserID(accountUsername)); //TODO could be better by username than by id?
             }
-
         }
-
-        //TODO add success or error banner
 
         System.out.println("Reservation has been added into DB");
 
     }
 
-    //todo cambiare uml
     public void sendInvites(Group group) throws SQLException, ClassNotFoundException {
 
         InviteSender inviteSender = new InviteSender(group);
@@ -139,7 +142,7 @@ public class UserActionsController extends PersonController{
         for (User user : receivers) {
             if(inviteDao.checkInvite(user.getId(),group.getId())){
                 System.out.println("Invite already exists");
-            }else{
+            }else if(user != null){
 
                 invite = inviteSender.factoryMethod();
                 invite.setUser(user);
@@ -176,7 +179,7 @@ public class UserActionsController extends PersonController{
     public ArrayList<WorkingHours> getWHsByFacilityByDay(int idFacility, DayOfWeek dayOfWeek) throws SQLException {
         WorkingHoursDAO workingHoursDAO = new WorkingHoursDAO();
 
-        return workingHoursDAO.getWHsByFacility(idFacility);
+        return workingHoursDAO.getWHsByFacilityByDay(idFacility,dayOfWeek);
     }
 
     public void declineInvite(int idInvite) throws SQLException {
@@ -186,12 +189,85 @@ public class UserActionsController extends PersonController{
 
     }
 
-    public void acceptInvite(Invite invite) throws SQLException, ClassNotFoundException {
-        //todo da aggiungere scelta guests
-        joinGroup(invite.getGroup().getId(), 0);
-        InviteDao inviteDao = new InviteDao();
+    //TODO changed to boolean (uml)
+    public boolean acceptInvite(Invite invite) throws SQLException, ClassNotFoundException {
 
-        inviteDao.deleteInvite(invite.getId());
+        boolean accepted = false;
+
+        if (invite.getGroup().getReservation().isMatched()) {
+
+            DialogPane selectGuestsDialogPane;
+            SelectGuestsPaneController selectGuestsPaneController;
+
+            //load guests selector
+            FXMLLoader loader = new FXMLLoader(getClass().getResource("/main/FXML/selectGuestsPane.fxml"));
+            try {
+                selectGuestsDialogPane = loader.load();
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+            selectGuestsPaneController = loader.getController(); //connect controller
+
+            Dialog<ButtonType> dialog = new Dialog<>();
+            dialog.setTitle("Who do you want to add?");
+            dialog.setDialogPane(selectGuestsDialogPane);
+
+            Button okButton = (Button) dialog.getDialogPane().lookupButton(ButtonType.OK);
+            okButton.setText("Send Invite");
+
+
+            okButton.addEventFilter(ActionEvent.ACTION, event -> {
+
+                int guests = selectGuestsPaneController.getnGuestsChoice().getValue() != null ? selectGuestsPaneController.getnGuestsChoice().getValue() : 0;
+                int accounts = selectGuestsPaneController.getAccountList().getItems().size();
+
+                boolean canJoin = invite.getGroup().canJoin(guests, accounts, true);
+
+                if (!canJoin) {
+                    event.consume(); // prevents dialog closing
+                    selectGuestsPaneController.getMessagesController().showMessage("Too much guests for this group.", MessagesController.MessageType.ERROR, 3);
+                }
+
+            });
+
+            Optional<ButtonType> result = dialog.showAndWait();
+
+            if (result.isPresent() && result.get() == ButtonType.OK) {
+
+                int guests = selectGuestsPaneController.getnGuestsChoice().getValue() != null ? selectGuestsPaneController.getnGuestsChoice().getValue() : 0;
+
+                //himself join into group
+                joinGroup(invite.getGroup().getId(), guests);
+
+                //send invites to other (his) players
+                ArrayList<String> accountsList = new ArrayList<>(selectGuestsPaneController.getAccountList().getItems());
+                UserDAO userDAO = new UserDAO();
+                for (String accountUsername : accountsList) {
+                    if (accountUsername != null) {
+                        sendInvite(invite.getGroup().getReservation(), userDAO.getUserID(accountUsername)); //TODO could be better by username than by id?
+                    }
+                }
+
+                accepted = true;
+
+                //delete this invite
+                InviteDao inviteDao = new InviteDao();
+                inviteDao.deleteInvite(invite.getId());
+
+            }
+
+
+        } else {
+            //guests are 0 because in not matched booking are not allowed guests
+            joinGroup(invite.getGroup().getId(), 0);
+            accepted = true;
+
+            //delete this invite
+            InviteDao inviteDao = new InviteDao();
+            inviteDao.deleteInvite(invite.getId());
+        }
+
+        return accepted;
     }
 
     public void joinGroup(int idGroup, int guestUsers) throws SQLException, ClassNotFoundException {
@@ -295,7 +371,7 @@ public class UserActionsController extends PersonController{
 
         IsPartDao isPartDao = new IsPartDao();
 
-        return isPartDao.getAllGroupsByUser(this.user.getId()); //FIXME id by method parameter or id like this?
+        return isPartDao.getAllGroupsByUser(this.user.getId());
 
     }
